% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.plot.R
\name{bs.plot}
\alias{bs.plot}
\title{Plots of Basis}
\usage{
bs.plot(type, k, or = 4, title = "")
}
\arguments{
\item{type}{type indicates which type of basis is used (There are 31 types in this package)}

\item{k}{The k is the number of basis functions represented (If wavelet are chosen, the real number of basis is 2^k. If
Cspli is chosen, the real number of basis is k-2+or)}

\item{or}{indicates the order of spline and only used in Cspli type, default is 4 which indicates cubic spline}

\item{title}{give the title for the basis plot}
}
\value{
The plot of 1 to k basis functions
}
\description{
bs.plot() generates the plot of first k basis function.
}
\examples{
\donttest{
bs.plot("Legen", 2)
bs.plot("tri", 3)
bs.plot("db10", 2)
}
}
