% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.test.R
\name{fix.test}
\alias{fix.test}
\title{The Test of Stability for AR Approximations With Fixed Parameters}
\usage{
fix.test(ts, c, b, type, or = 4, B.s = 1000, m = 0)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{c indicates the number of basis used to estimate (For wavelet, the number of basis is 2^c. If
Cspli is chosen, the real number of basis is c-2+or)}

\item{b}{b is the lag for auto-regressive model}

\item{type}{type indicates which type of basis is used. There are 31 types in this package}

\item{or}{indicates the order of spline and only used in Cspli type, default is 4 which indicates cubic spline}

\item{B.s}{the number of statistics used in multiplier bootstrap, the default value is 1000}

\item{m}{the number of window size used in multiplier bootstrap, the default value is 0 which uses the
minimum volatility method to determine the number}
}
\value{
p value of the test
}
\description{
fix.test() generates a test of Stability for AR Approximations with fixed parameters.
}
\examples{
\donttest{
set.seed(137)
time.series = c()
n = 1024
v = 25
w = rnorm(n, 0, 1) / v
x_ini = runif(1,0,1)
for(i in 1:n){
  if(i == 1){
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*x_ini  + w[i] #
  } else{
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*time.series[i-1] + w[i]
  }
}
fix.test(time.series, 5, 1, type = "Legen")
fix.test(time.series, 3, 1, type = "db10")
}
}
