% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie.auto.pacf.R
\name{sie.auto.pacf}
\alias{sie.auto.pacf}
\title{Generate Pacf Automatically}
\usage{
sie.auto.pacf(ts, c, lag, type, or = 4, m = 500)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{c indicates the number of basis used to estimate (For wavelet, the number of basis is 2^c.If
Cspli is chosen, the real number of basis is c-2+or)}

\item{lag}{lag b is the lag for auto-regressive model}

\item{type}{type indicates which type of basis is used (There are 31 types in this package)}

\item{or}{indicates the order of spline and only used in Cspli type, default is 4 which indicates cubic spline}

\item{m}{m indicates the number of points of coefficients to estimate}
}
\value{
A list contains the PACF in each lag
}
\description{
sie.auto.pacf() generates the PACF from 1 to lag automatically.
}
\examples{
\donttest{
set.seed(137)
time.series = c()
n = 1024
v = 25
w = rnorm(n, 0, 1) / v
x_ini = runif(1,0,1)
for(i in 1:n){
  if(i == 1){
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*x_ini  + w[i] #
  } else{
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*time.series[i-1] + w[i]
  }
}
sie.auto.pacf(time.series, 5, 1, "Legen")
sie.auto.pacf(time.series, 2, 1, "db10", 600)
}
}
