% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie.auto.plot.R
\name{sie.auto.plot}
\alias{sie.auto.plot}
\title{Plot the Estimate Results by Automatic Fitting}
\usage{
sie.auto.plot(res.auto.fit, type, title = "")
}
\arguments{
\item{res.auto.fit}{the output from sie.auto.fit() function}

\item{type}{type indicates which type of basis is used (There are 31 types in this package)}

\item{title}{give the title for the auto estimate plot}
}
\value{
A list which contains 3 plot related to the estimation of coefficient, Elbow point and cross validation in order
}
\description{
sie.auto.plot() visualizes the estimation of coefficient, gives the elbow plot and represents the cross validation result.
}
