% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie.plot.pacf.R
\name{sie.plot.pacf}
\alias{sie.plot.pacf}
\title{Plot Pacf}
\usage{
sie.plot.pacf(ts, c, lag, type, ops = "2d", title = "", m = 500, or = 4)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{c indicates the number of basis used to estimate (For wavelet, the number of basis is 2^c. If
Cspli is chosen, the real number of basis is c-2+or)}

\item{lag}{lag b is the lag for auto-regressive model}

\item{type}{type indicates which type of basis is used (There are 32 types in this package)}

\item{ops}{choose 2D plot or 3D plot ("2d" inicates 2D plot and "3d" indicates 3D plot)}

\item{title}{give the title for the pacf plot}

\item{m}{m indicates the number of points of coefficients to estimate}

\item{or}{or indicates the order of spline, default is 4 which indicates cubic spline}
}
\value{
The plot of pacf basis on the time series data
}
\description{
sie.plot.pacf() shows the PACF with different lag.
}
\examples{
\donttest{
set.seed(137)
time.series = c()
n = 1024
v = 25
w = rnorm(n, 0, 1) / v
x_ini = runif(1,0,1)
for(i in 1:n){
  if(i == 1){
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*x_ini  + w[i] #
  } else{
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*time.series[i-1] + w[i]
  }
}

sie.plot.pacf(time.series, 5, 10, "Legen")
sie.plot.pacf(time.series, 5, 10, "Legen", "3d")
}
}
