% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.pacf.test.R
\name{fix.pacf.test}
\alias{fix.pacf.test}
\title{Testing Lag of Auto-Regressive (AR) Model}
\usage{
fix.pacf.test(ts, c, type, or = 4, lag = 3, b = 8, B.s = 1000, m = 0)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{c indicates the number of basis used to estimate (For wavelet, the number of basis is 2^c. If
Cspli is chosen, the real number of basis is c-2+or)}

\item{type}{type indicates which type of basis is used. There are 31 types in this package}

\item{or}{or indicates the order of spline and only used in Cspli type, default is 4 which indicates cubic spline}

\item{lag}{lag determine the lag of AR Approximations.The default is 3}

\item{b}{the largest lag for auto-regressive model, the default value is 8, this parameter must be
larger than lag}

\item{B.s}{the number of statistics used in multiplier bootstrap, the default value is 1000}

\item{m}{the number of window size used in multiplier bootstrap, the default value is 0 which uses the
minimum volatility method to determine the number}
}
\value{
It returns a list contains p value for each lag
}
\description{
fix.pacf.test() generates a test of lags for AR Approximations.
}
