% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie.auto.plot.R
\name{sie.auto.plot}
\alias{sie.auto.plot}
\title{Plot the Estimate Results by Automatic Fitting}
\usage{
sie.auto.plot(res.auto.fit, type, title = "")
}
\arguments{
\item{res.auto.fit}{the output from sie.auto.fit() function}

\item{type}{type indicates which type of basis is used (There are 31 types in this package)}

\item{title}{give the title for the auto estimate plot}
}
\value{
A list which contains 3 plot related to the estimation of coefficient, Elbow point and cross validation in order
}
\description{
sie.auto.plot() visualizes the estimation of coefficient, gives the elbow plot and represents the cross validation result.
}
\examples{
\donttest{
set.seed(137)
time.series = c()
n = 1024
v = 25
w = rnorm(n, 0, 1) / v
x_ini = runif(1,0,1)
for(i in 1:n){
  if(i == 1){
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*x_ini  + w[i] #
  } else{
    time.series[i] = 0.2 + 0.6*cos(2*pi*(i/n))*time.series[i-1] + w[i]
  }
}

res1 = sie.auto.fit(time.series, type = "db10")
sie.auto.plot(res1, "db10")
res2 = sie.auto.fit(time.series, type = "Legen", method = "LOOCV")
sie.auto.plot(res2, "Legen")
}
}
