% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sie.plot.pacf.R
\name{sie.plot.pacf}
\alias{sie.plot.pacf}
\title{Plot Partial Autocorrelation Function (PACF)}
\usage{
sie.plot.pacf(
  ts,
  c,
  lag,
  type,
  ops = "2d",
  title = "",
  m = 500,
  or = 4,
  se = FALSE
)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{c indicates the number of basis used to estimate (For wavelet, the number of basis is 2^c. If
Cspli is chosen, the real number of basis is c-2+or)}

\item{lag}{lag b is the lag for auto-regressive model}

\item{type}{type indicates which type of basis is used (There are 32 types in this package)}

\item{ops}{choose 2D plot or 3D plot ("2d" inicates 2D plot and "3d" indicates 3D plot)}

\item{title}{give the title for the pacf plot}

\item{m}{m indicates the number of points of coefficients to estimate}

\item{or}{or indicates the order of spline, default is 4 which indicates cubic spline}

\item{se}{The option to indicate whether the 95\% confidence interval is displayed or not, the default is FALSE}
}
\value{
The plot of pacf basis on the time series data
}
\description{
sie.plot.pacf() shows the PACF with different lag.
}
